+++ src/business/entities/Inventory.ts
@@ -1,0 +1,8 @@
+import { Warehouse } from "./Warehouse"
+
+
+
+export interface Inventory {
+    quantity?: number;
+    warehouses: Warehouse[];
+}
