+++ src/controller/model/Inventory.ts
@@ -1,0 +1,6 @@
+import { WarehouseInterface } from "./Warehouse";
+
+export interface InventoryInterface {
+    quantity?: number;
+    warehouses: WarehouseInterface[];
+}
