+++ src/controller/model/Product.ts
@@ -1,0 +1,8 @@
+import { InventoryInterface } from "./Inventory";
+
+
+export interface ProductInputDTO {
+    sku: number;
+    name: string;
+    inventory: InventoryInterface;
+}
