+++ src/controller/model/Warehouse.ts
@@ -1,0 +1,5 @@
+export interface WarehouseInterface {
+    locality: string;
+    quantity: number;
+    type: string;
+}
