+++ src/controller/routes/ProductRouter.ts
@@ -1,11 +1,12 @@
 import express from "express";
-import { ProductController } from "../controller/ProductController";
+import { ProductController } from "../ProductController";
 
 export const productRouter = express.Router()
 
 const productController = new ProductController()
 
 productRouter.post("/create", productController.createProduct)
+productRouter.put("/edit/:sku", productController.editProduct)
 productRouter.delete("/delete/:sku", productController.deleteProduct)
 productRouter.get("/:sku", productController.getProductBySku)
